pyproject.toml
[project]
name = "jazzevent"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "oleksiymosiychuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "pip (>=25.3,<26.0)",
    "django-phonenumber-field[phonenumbers] (>=8.3.0,<9.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "django-vite (>=3.1.0,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "sorl-thumbnail (>=13.0.0,<14.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-admin-sortable2 (>=2.3.1,<3.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^6.2.1"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119